+++ tests/files/src/0500_specific_property.php
@@ -1,0 +1,21 @@
+<?php
+
+class Example500 {
+    /** @var string[] */
+    private $propWithArrayDefault = [];
+
+    /** @var string[] */
+    private $propSetInConstructor;
+
+    public function __construct() {
+        $this->propSetInConstructor = [];
+    }
+
+    public function badAssign(int $val) {
+        $this->propWithArrayDefault[] = $val;
+    }
+
+    public function badAssign2(int $val) {
+        $this->propSetInConstructor[] = $val;
+    }
+}